backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.8
SQLAlchemy>=1.4.17,<=1.4.41
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings
python-dotenv==1.0.0
google-auth==2.23.4
requests==2.31.0